[Insights] Add aggregation to the Size, StartTime and EndTime columns of AllocTreeView. #rb none
[CL 14989497 by Catalin Dragoiu in ue5-main branch]

+++ Engine/Source/Developer/TraceInsights/Private/Insights/Table/ViewModels/TableCellValue.h
@@ -115,7 +115,7 @@ public:
 		}
 	}
 
-	float AsDouble() const
+	double AsDouble() const
 	{
 		switch (DataType)
 		{
